This commit was manufactured by cvs2svn to create tag 'SpecTcl-2-2-ENDOFLIFE'.

+++ SpecTcl/Display/dispgrob.cc
@@ -662,7 +662,7 @@ void grobj_sum1d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -734,7 +734,7 @@ void grobj_cut1d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -806,7 +806,7 @@ void grobj_mark1d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -880,7 +880,7 @@ void grobj_contour::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -959,7 +959,7 @@ void grobj_band::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -1035,7 +1035,7 @@ void grobj_sum2d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -1114,7 +1114,7 @@ void grobj_mark2d::draw(XMWidget *pane, win_attributed *at, Boolean final)
   /* Determine if there's a backing store too.. we rely on the fact that
   ** the pane has encoded in it the widget index:
   */
-  int index, row, col;
+  long index, row, col;
   pane->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
@@ -1276,7 +1276,7 @@ void grobj_ptlist1::draw(XMWidget *p, win_attributed *at, Boolean final)
   **  so that if the window is contaiminated while accepting points, the 
   **  partial point set will be visible.
   */
-  int index, row, col;
+  long index, row, col;
   p->GetAttribute(XmNuserData, &index); // This is the pane number.
   row = index % WINDOW_MAXAXIS;
   col = index / WINDOW_MAXAXIS;
@@ -1410,7 +1410,7 @@ void grobj_ptlist2::draw(XMWidget *p, win_attributed *at, Boolean final)
   **  so that if the window is contaiminated while accepting points, the 
   **  partial point set will be visible.
   */
-  int index, row, col;
+  long index, row, col;
   p->GetAttribute(XmNuserData, &index); // This is the pane number.
   row = index % WINDOW_MAXAXIS;
   col = index / WINDOW_MAXAXIS;
@@ -1514,7 +1514,7 @@ void grobj_Peak1d::draw(XMWidget* pWindow,
   // we rely on the fact that the pane widget has the pane number in it's
   // user data field:
 
-  int index, row, col;
+  long index, row, col;
   pWindow->GetAttribute(XmNuserData, &index);
   col = index % WINDOW_MAXAXIS;
   row = index / WINDOW_MAXAXIS;
